Moves Tests and Reformatted
Renamed attr_collecter to collect_book_attrs.

Moved tests for collect_book_attrs to test_services.py.

Changed the post_init in book dataclass to always apply the "...".
